requirements.txt
psycopg2-binary==2.9
git+https://github.com/moveonorg/pywell@main#egg=pywell
requests==2.22.0
slackweb==1.0.5
boto3>=1.20.24
botocore>=1.23.24
